# server/openapi.yaml
openapi: 3.0.3
info:
  title: Aichi Governance API
  version: 1.0.0
servers:
  - url: /
paths:
...
  /api/v1/reviews:
    get:
      summary: List reviews
      parameters:
        - in: query
          name: status
          schema:
            type: string
      responses:
        '200':
          description: Review list
    post:
      summary: Create review draft
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, type]
      responses:
        '201':
          description: Review created
...